msr: add mwait and residency support
When using mwait, you may specify a desired C* state, how deep the CPU
may go to sleep.

The commits adds support to report mwait support and to enable mwait in the
nova kernel. Msr accepts the C* hint which will be forwarded to the NOVA
kernel.

Additionally, the commit adds support to read out Core and Package residency
counter provided by the kernel, which reflects the time how long the CPU was
in which C* state.